+++ app/Http/Controllers/ButirSikapController.php
@@ -8,9 +8,14 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Response;
 use Excel;
+use Illuminate\Support\Facades\Auth;
 
 class ButirSikapController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['auth','revalidate']);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -18,9 +23,13 @@ class ButirSikapController extends Controller
      */
     public function index()
     {
-        $title = 'Butir-Butir Sikap';
-        $data_sikap = ButirSikap::orderBy('kategori_butir_id', 'ASC')->get();
-        return view('admin.butir-sikap.index', compact('title', 'data_sikap'));
+        if(Auth::user()->hasRole('admin')){
+            $title = 'Butir-Butir Sikap';
+            $data_sikap = ButirSikap::orderBy('kategori_butir_id', 'ASC')->get();
+            return view('admin.butir-sikap.index', compact('title', 'data_sikap'));
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
+        }
     }
 
     /**
@@ -30,11 +39,15 @@ class ButirSikapController extends Controller
      */
     public function create()
     {
-        $file = public_path() . "/format_excel/format_import_sikap.xlsx";
-        $headers = array(
-            'Content-Type: application/xlsx',
-        );
-        return Response::download($file, 'format_import_sikap ' . date('Y-m-d H_i_s') . '.xlsx', $headers);
+        if(Auth::user()->hasRole('admin')){
+            $file = public_path() . "/format_excel/format_import_sikap.xlsx";
+            $headers = array(
+                'Content-Type: application/xlsx',
+            );
+            return Response::download($file, 'format_import_sikap ' . date('Y-m-d H_i_s') . '.xlsx', $headers);
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
+        }
     }
 
     /**
@@ -45,21 +58,25 @@ class ButirSikapController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'kategori_butir_id' => 'required',
-            'kode' => 'required|min:2|max:10|unique:butir_sikap',
-            'butir_sikap' => 'required|min:4|max:255',
-        ]);
-        if ($validator->fails()) {
-            return back()->with('error', $validator->messages()->all()[0])->withInput();
-        } else {
-            $sikap = new ButirSikap([
-                'kategori_butir_id' => $request->jenis_kompetensi,
-                'kode' => $request->kode,
-                'butir_sikap' => $request->butir_sikap,
+        if(Auth::user()->hasRole('admin')){
+            $validator = Validator::make($request->all(), [
+                'kategori_butir_id' => 'required',
+                'kode' => 'required|min:2|max:10|unique:butir_sikap',
+                'butir_sikap' => 'required|min:4|max:255',
             ]);
-            $sikap->save();
-            return back()->with('success', 'Butir sikap berhasil ditambahkan');
+            if ($validator->fails()) {
+                return back()->with('error', $validator->messages()->all()[0])->withInput();
+            } else {
+                $sikap = new ButirSikap([
+                    'kategori_butir_id' => $request->jenis_kompetensi,
+                    'kode' => $request->kode,
+                    'butir_sikap' => $request->butir_sikap,
+                ]);
+                $sikap->save();
+                return back()->with('success', 'Butir sikap berhasil ditambahkan');
+            }
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
         }
     }
 
@@ -94,24 +111,28 @@ class ButirSikapController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validator = Validator::make($request->all(), [
-            'kode' => 'required|min:2|max:10|unique:butir_sikap' . ($id ? ",id,$id" : ''),
-            'butir_sikap' => 'required|min:4|max:255',
-        ]);
-        if ($validator->fails()) {
-            return back()->with('error', $validator->messages()->all()[0])->withInput();
-        } else {
-            try {
-                $sikap = ButirSikap::findorfail($id);
-                $data_sikap = [
-                    'kode' => $request->kode,
-                    'butir_sikap' => $request->butir_sikap,
-                ];
-                $sikap->update($data_sikap);
-                return back()->with('success', 'Butir sikap berhasil diedit');
-            } catch (Exception $e) {
-                return back()->with('error', 'kode sudah ada sebelumnya');
+        if(Auth::user()->hasRole('admin')){
+            $validator = Validator::make($request->all(), [
+                'kode' => 'required|min:2|max:10|unique:butir_sikap' . ($id ? ",id,$id" : ''),
+                'butir_sikap' => 'required|min:4|max:255',
+            ]);
+            if ($validator->fails()) {
+                return back()->with('error', $validator->messages()->all()[0])->withInput();
+            } else {
+                try {
+                    $sikap = ButirSikap::findorfail($id);
+                    $data_sikap = [
+                        'kode' => $request->kode,
+                        'butir_sikap' => $request->butir_sikap,
+                    ];
+                    $sikap->update($data_sikap);
+                    return back()->with('success', 'Butir sikap berhasil diedit');
+                } catch (Exception $e) {
+                    return back()->with('error', 'kode sudah ada sebelumnya');
+                }
             }
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
         }
     }
 
@@ -123,18 +144,26 @@ class ButirSikapController extends Controller
      */
     public function destroy($id)
     {
-        $sikap = ButirSikap::findorfail($id);
-        $sikap->delete();
-        return back()->with('success', 'Sukses! Tahun Pelajaran Dihapus');
+        if(Auth::user()->hasRole('admin')){
+            $sikap = ButirSikap::findorfail($id);
+            $sikap->delete();
+            return back()->with('success', 'Sukses! Tahun Pelajaran Dihapus');
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
+        }
     }
 
     public function import(Request $request)
     {
-        try {
-            Excel::import(new ButirSikapImport, $request->file('file_import'));
-            return back()->with('success', 'Data butir sikap berhasil diimport');
-        } catch (Exception $e) {
-            return back()->with('error', 'Maaf, format data tidak sesuai');
+        if(Auth::user()->hasRole('admin')){
+            try {
+                Excel::import(new ButirSikapImport, $request->file('file_import'));
+                return back()->with('success', 'Data butir sikap berhasil diimport');
+            } catch (Exception $e) {
+                return back()->with('error', 'Maaf, format data tidak sesuai');
+            }
+        }else{
+            return response()->view('errors.403', [abort(403), 403]);
         }
     }
 }
